+++ lib/transpec/syntax/raise_error.rb
@@ -20,8 +20,13 @@ module Transpec
       end
 
       def positive?
-        expectation_method_name = parent_node.children[1]
-        [:should, :to].include?(expectation_method_name)
+        ancestor_nodes.reverse_each do |ancestor_node|
+          next unless ancestor_node.type == :send
+          expectation_method_name = ancestor_node.children[1]
+          return [:should, :to].include?(expectation_method_name)
+        end
+
+        false
       end
 
       private